HBX-1989: Move class 'PrimaryKeyProcessor' from 'org.hibernate.tool.internal.reveng' to 'org.hibernate.tool.internal.reveng.reader'
